Bidirectional Dictionary added
this patch adds a bidirectional dictionary query. Now you have to write
the language-direction only for vocabularies, the dictionary
automatically finds expressions in both l1 and l2 at one singular query.
You write "g" and the ouput is:
Results:

git=genial
genial=git

Found 2 matches

Well, in this example it doesn't make sense but else it is a handy
feature.