Random order added.
If you call the script with a "r" at the end of the list of params, it'll ask you the words in random order. If you type instead of "r" an "o", it'll ask you the words in the old, linear order (1,2,3...) In the next time, I write the Documentation.