Write mode added
Added the write mode. With this mode you can create (but not edit yet) the vocabulary db files. This is the first and raw implementation, it's to be enhanced!!!
It asks you almost like the training mode for word pairs, collects them in an array and writes them finally into the file given with the first argument. Attention: It overwrites the file! This
feature is to be documentated and enhanced. (Editing mode?)